+++ src/main/kotlin/com/example/kbocombo/combo/ui/ComboController.kt
@@ -1,7 +1,7 @@
 package com.example.kbocombo.combo.ui
 
 import com.example.kbocombo.combo.application.ComboService
-import com.example.kbocombo.combo.ui.request.ComboCreateRequest
+import com.example.kbocombo.combo.application.request.ComboCreateRequest
 import com.example.kbocombo.member.domain.Member
 import com.example.kbocombo.member.ui.MemberResolver
 import org.springframework.web.bind.annotation.DeleteMapping
